+++ app/src/main/java/com/hefengbao/jingmo/data/repository/IdiomRepository.kt
@@ -5,11 +5,11 @@ import com.hefengbao.jingmo.data.database.model.SimpleIdiomInfo
 import kotlinx.coroutines.flow.Flow
 
 interface IdiomRepository {
-    suspend fun getIdiom(id: Long): IdiomEntity
-    suspend fun getNextId(id: Long): Long
-    suspend fun getPrevId(id: Long): Long
+    suspend fun getIdiom(id: Int): IdiomEntity
+    suspend fun getNextId(id: Int): Int
+    suspend fun getPrevId(id: Int): Int
     fun getSimpleIdiomInfoList(): Flow<List<SimpleIdiomInfo>>
     fun searchSimpleIdiomInfoList(query: String): Flow<List<SimpleIdiomInfo>>
-    suspend fun getSearchNextId(id: Long, query: String): Long
-    suspend fun getSearchPrevId(id: Long, query: String): Long
+    suspend fun getSearchNextId(id: Int, query: String): Int
+    suspend fun getSearchPrevId(id: Int, query: String): Int
 }
