+++ app/src/main/java/com/hefengbao/jingmo/data/repository/IdiomRepositoryImpl.kt
@@ -9,18 +9,18 @@ import javax.inject.Inject
 class IdiomRepositoryImpl @Inject constructor(
     private val idiomDao: IdiomDao
 ) : IdiomRepository {
-    override suspend fun getIdiom(id: Long): IdiomEntity = idiomDao.getIdiom(id)
-    override suspend fun getNextId(id: Long): Long = idiomDao.getNextId(id)
-    override suspend fun getPrevId(id: Long): Long = idiomDao.getPrevId(id)
+    override suspend fun getIdiom(id: Int): IdiomEntity = idiomDao.getIdiom(id)
+    override suspend fun getNextId(id: Int): Int = idiomDao.getNextId(id)
+    override suspend fun getPrevId(id: Int): Int = idiomDao.getPrevId(id)
     override fun getSimpleIdiomInfoList(): Flow<List<SimpleIdiomInfo>> =
         idiomDao.getSimpleIdiomInfoList()
 
     override fun searchSimpleIdiomInfoList(query: String): Flow<List<SimpleIdiomInfo>> =
         idiomDao.searchSimpleIdiomInfoList("%$query%")
 
-    override suspend fun getSearchNextId(id: Long, query: String): Long =
+    override suspend fun getSearchNextId(id: Int, query: String): Int =
         idiomDao.getSearchNextId(id, "%$query%")
 
-    override suspend fun getSearchPrevId(id: Long, query: String): Long =
+    override suspend fun getSearchPrevId(id: Int, query: String): Int =
         idiomDao.getSearchPrevId(id, "%$query%")
 }
